// frontend/src/shared/api/admin/model/listReportsStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * defShows Admin API
 * OpenAPI spec version: 0.1.0
 */

export type ListReportsStatus = typeof ListReportsStatus[keyof typeof ListReportsStatus];


export const ListReportsStatus = {
  open: 'open',
  resolved: 'resolved',
  dismissed: 'dismissed',
} as const;