// frontend/src/shared/api/admin/model/report.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * defShows Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportReason } from './reportReason';
import type { ReportStatus } from './reportStatus';
import type { ReportUserSummary } from './reportUserSummary';

export interface Report {
  id: number;
  reporter: ReportUserSummary;
  target: ReportUserSummary;
  reason: ReportReason;
  note: string;
  status: ReportStatus;
  created_at: string;
  resolved_at?: string;
}