// frontend/src/shared/api/admin/model/reportReason.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * defShows Admin API
 * OpenAPI spec version: 0.1.0
 */

export type ReportReason = typeof ReportReason[keyof typeof ReportReason];


export const ReportReason = {
  spam: 'spam',
  harassment: 'harassment',
  inappropriate: 'inappropriate',
  other: 'other',
} as const;